requirements.txt
apache-airflow[celery, redis, microsoft.mssql, odbc, common.io, common.sql, async, pandas, ftp]==2.9.1
pyarrow==16.0.0
psycopg2-binary==2.9.9
ipywidgets==8.1.2
openpyxl==3.1.2
